fix: maintenance windows security, DoS protection, and wire-format compatibility
Address findings from the AI-reviewer review report on the maintenance windows PR.

Security (Critical):
- Remove maintenance exemption from BeforeValidatorSlashed collateral hook. The SDK's staking.Slash() invokes this hook for ALL slash reasons, including double-sign / equivocation. The previous unconditional skip during active maintenance allowed an attacker to schedule a window and double-sign without losing collateral. Downtime exemption remains enforced upstream in the SDK fork's x/slashing liveness path; punitive slashes that reach this hook must be applied.

DoS protection:
- Introduce MaintenanceScheduledIndex (KeySet, prefix 69) maintained on schedule/cancel/activate. Replaces full MaintenanceStates iteration in collectActiveAndScheduledReservations with bounded scans of the active + scheduled indexes — query cost drops from O(participants) to O(actual maintenance count). Public concurrency / schedulability queries can no longer be used to burn unbounded CPU/IO at zero gas cost.
- Extract iterateIndexedReservations helper that surfaces iter.Key() errors via LogError instead of silently continuing, and enforces a hard maxMaintenanceIterationLimit = 10000 safety bound on every index walk.
- MaintenanceActive query and CollectActiveMaintenanceAddresses both route through the helper.

Correctness:
- Fix off-by-one in lead-time validation (msg_server_schedule_maintenance.go and query_maintenance.go): change `<=` to `<` so a request scheduled exactly MinScheduleLeadBlocks blocks ahead is accepted.
- Always delete a transition row in ProcessMaintenanceTransitions, even when the activate/complete handler returns an error. Previously failed rows leaked storage permanently (height-keyed entries are never revisited); failures are surfaced via LogError and rely on operator monitoring rather than infinite on-chain retry.
- Document the existing idempotency invariant of GrantMaintenanceCredit: finishSettle removes the SettleAmount before crediting, and the whole flow runs inside a cached context committed atomically with the claim, so a duplicate call within the same epoch cannot reach the grant path.

Wire-format compatibility:
- Swap proto field numbers so delegation_params stays at 16 (matching the upstream multi-model branch that shipped first) and maintenance_params becomes 17. Renumbering would silently lose DelegationParams from any chain that already stored params under field 16. Generated params.pb.go updated in lockstep (struct tags, Unmarshal case 16/17, Marshal write order, and wire-tag bytes 0x82,0x01 / 0x8a,0x01).

Logging:
- Demote 8 LogInfo calls in GetRandomExecutor / createFilterFn / createIsAvailableDuringPoCFilterFn to LogDebug. These fired on every executor selection request and produced excessive noise on a live chain. Drop the timeslot_allocation field from the per-node "available during PoC" log line.

Test improvements (Go):
- Replace 21 hardcoded gonka1... bech32 literals in maintenance_test.go with sample.AccAddress(). Tests no longer break if the bech32 prefix changes or addresses are regenerated.
- Add `tc := tc` capture in both table-driven t.Parallel loops (TestScheduleMaintenance_Failures, TestSchedulability_Failures) to avoid loop-variable aliasing across parallel subtests.

Test improvements (Kotlin):
- Widen MaintenanceParams.maintenanceMaxConcurrentValidators and maintenanceMaxConcurrentPowerBps from Int to Long (proto type is uint32, range exceeds Kotlin Int.MAX_VALUE). Update MaintenanceWindowTests.kt literals with L suffix.

+++ testermint/src/main/kotlin/data/AppExport.kt
@@ -523,10 +523,13 @@ data class MaintenanceParams(
     val maintenanceMinScheduleLeadBlocks: Long = 100,
     @SerializedName("maintenance_max_window_blocks")
     val maintenanceMaxWindowBlocks: Long = 200,
+    // Proto types are uint32 (range 0 .. 4_294_967_295). Kotlin Int is signed
+    // and would overflow at 2_147_483_648. Widen to Long so any governance
+    // value the chain accepts can round-trip without silent truncation.
     @SerializedName("maintenance_max_concurrent_validators")
-    val maintenanceMaxConcurrentValidators: Int = 3,
+    val maintenanceMaxConcurrentValidators: Long = 3,
     @SerializedName("maintenance_max_concurrent_power_bps")
-    val maintenanceMaxConcurrentPowerBps: Int = 1000,
+    val maintenanceMaxConcurrentPowerBps: Long = 1000,
     @SerializedName("maintenance_credit_cap_blocks")
     val maintenanceCreditCapBlocks: Long = 400,
     @SerializedName("maintenance_credit_earn_per_successful_epoch_blocks")
